requirements.txt
aiosqlite==0.21.0
alembic==1.16.4
annotated-types==0.7.0
anyio==4.10.0
apprise==1.9.4
asgi-lifespan==2.1.0
asyncpg==0.30.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.1
colorama==0.4.6
coolname==2.2.0
cryptography==45.0.6
dateparser==1.2.2
decorator==5.2.1
docker==7.1.0
exceptiongroup==1.3.0
fastapi==0.116.1
Flask==3.1.0
fsspec==2025.7.0
future==1.0.0
geocoder==1.38.1
google-api-core==2.23.0
google-api-python-client==2.152.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
graphviz==0.21
greenlet==3.2.4
griffe==1.11.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jq==1.9.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.1
oauthlib==3.2.2
opensky-api @ file:///Users/admin/code/Aviation_postgres_prefect/opensky-api/python
opentelemetry-api==1.36.0
orjson==3.11.1
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
prefect==3.4.12
prometheus_client==0.22.1
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
python-socks==2.7.2
pytz==2024.2
PyYAML==6.0.2
ratelim==0.1.6
readchar==4.2.1
referencing==0.36.2
regex==2025.7.34
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.1.0
rpds-py==0.27.0
rsa==4.9
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
semver==3.0.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.47.2
text-unidecode==1.3
toml==0.10.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.2.3
uv==0.8.8
uvicorn==0.35.0
websockets==15.0.1
Werkzeug==3.1.3
wget==3.2
whenever==0.8.8
zipp==3.23.0
